requirements.txt
pyyaml>=5.3.1
requests>=2.22.0